+++ src/Algorithm/Arrays/LeetcodeProblems/Find_Kth_Largest/Find_Kth_Largest.cpp
@@ -1,0 +1,10 @@
+class Solution {
+public:
+    int findKthLargest(vector<int>& nums, int k) {
+        int len = nums.size();
+        // sort the array
+        sort(nums.begin(), nums.end());
+        // return kth largest
+        return nums[len-k];
+    }
+};
